iwd: Stop using _nm_utils_ssid_to_utf8()
_nm_utils_ssid_to_utf8() can be quite heavy and also has this comment:

 * Again, this function should be used for debugging and display purposes
 * _only_.

In most places that we used it, we have already validated the
connection's SSID to be valid UTF-8 so we can simply g_strndup() it now,
even in the two places where we actually only needed it for display
purposes.  And we definitely don't need or want the locale-specific
conversions done in _nm_utils_ssid_to_utf8 when the SSID is *not* utf8.

In mirror_8021x_connection we also optimize the lookup loop to avoid
validating and strdup'ing all the SSID.