iwd: Add the wifi.iwd.autoconnect setting
If this setting it true (or missing) we skip most of the D-Bus
Disconnect() calls whoe purpose was to keep IWD's internal autoconnect
mechanism always disabled.  We use the IWD's Station.State property
updates, and secrets requets through our IWD agent, to find out when IWD
is trying to connect and create "assumed" activations on the NM side to
mirror the IWD state.  This is quite complicated due to the many
possible combinations of NMDevice's state and IWD's state.  A lot of
them are "impossible" but we try to be careful to consider all the
different possibilities.

NM has a nice API for "assuming" connections, I'm not sure if that was
designed only for when NM starts or also when a connection is made
"under" it dynamically, but in any case this *could* be used for our use
case here.  However there were a few minor reasons I didn't want to use
it, which are listed in the comment in assume_connection().  Those can
be fixed or improved in NMManager / NMDevice and NMDeviceIiwd could then
start using that API maybe, though there's really not much difference
for NMDeviceIwd if that API was used.  For now I think it's fine if we
create a normal "managed"-type connection.